Update README.md
The initialism is 'VM', and we don't need an apostrophe when pluralising.  It's (apparently) also ESXi, OVFTool and Windows (but respect variable names).   The abbreviation for infrastructure could be 'infra', but isn't 'infa'; and it's so hard to type we could maybe just use i11e?  That would be a great thing, right?   :-D